Order pt writes before jumping to supervisor mode
Although, usually, caches for in-memory memory-management data
structures (e.g. TLB) are empty on reset, implicit references to such
data structures, due to instruction execution, might not necessarily be
ordered with respect to explicit writes. An SFENCE.VMA guarantees such
ordering.

This is necessary in order to ensure that when trapping into S mode the
hart sees the correct address translation mappings for supervisor level
code. For user level code, the supervisor already executes an SFENCE.VMA
when mapping the pages.