Fix cputchar on RV32 with virtual memory
When cputchar was fixed for RV32, virtual memory was not taken into
consideration. If cputchar was called in supervisor mode, with virtual
memory enabled, the pointers given to the HTIF would have wrongfully
been virtual addresses.

Translate pointers to always be physical addresses to support cputchar
with both virtual memory enabled and disabled. Defined macro kaa2pa is
short for 'kernel any address to physical address'.